+++ shared/types.ts
@@ -2,6 +2,16 @@ import { MessageParam } from '@anthropic-ai/sdk/resources/index.mjs'
 
 export type ModelDriverType = 'anthropic' | 'ollama' | 'openai'
 
+export type SignalType = 'cron' | 'state' | 'event'
+
+export type AutomationType = 'manual' | 'determine-signal' | 'execute-signal'
+
+export interface Automation {
+  hash: string
+  contents: string
+  fileName: string
+}
+
 export type ScenarioResult = {
   prompt: string
   toolsDescription: string
